temp-sculpt-normals-masking: Replace a few usages of float[3] with float3

+++ source/blender/blenkernel/BKE_paint.h
@@ -747,7 +747,7 @@ typedef struct SculptSession {
   char *last_paint_canvas_key;
 
   float last_normal[3];
-  int stroke_id;
+  uchar stroke_id;
 } SculptSession;
 
 void BKE_sculptsession_free(struct Object *ob);
